==> src/1_metaFix.Rmd <==
---
title: "Metadata curation"
author: "Ece Kartal"
date: "5.10.2022"
output:
  pdf_document: default
  html_document: default
---

Metadata is in several excel sheet, we first combine them to work with it.
```{r}
# Set parameters
PARAM <- list()
PARAM$folder.R <- paste0(getwd(), "/")
PARAM$folder <- gsub("src/", "", PARAM$folder.R)
PARAM$folder.input <- paste0(PARAM$folder, "input/")
PARAM$folder.Rdata <- paste0(PARAM$folder, "Rdata/")

library(tidyverse)
library(readxl)
```

```{r}
# read meta files 
Metadata_PROMISE <- read_excel(paste0(PARAM$folder.input,"/metadata/clean/Metadata_PROMISE.xlsx"))
meta1 <- read_excel(paste0(PARAM$folder.input, "metadata/Summary_vital_status_PROMISE_20231001.xlsx"), sheet = "Vital_status")
meta2 <-  read_excel(paste0(PARAM$folder.input, "metadata/Summary_vital_status_PROMISE_20231001.xlsx"), sheet = "Therapy")
meta3 <-  read_excel(paste0(PARAM$folder.input, "metadata/Summary_vital_status_PROMISE_20231001.xlsx"), sheet = "Response")

## TODO ##
meta4 <-  read_excel(paste0(PARAM$folder.input, "metadata/Summary_vital_status_PROMISE_20231001.xlsx"), sheet = "RECIST")
meta5 <-  read_excel(paste0(PARAM$folder.input, "metadata/Summary_vital_status_PROMISE_20231001.xlsx"), sheet = "weight_bsa_cycle")
meta6 <-  read_excel(paste0(PARAM$folder.input, "metadata/Summary_vital_status_PROMISE_20231001.xlsx"), sheet = "Antibiotics")
meta7 <-  read_excel(paste0(PARAM$folder.input, "metadata/Summary_vital_status_PROMISE_20231001.xlsx"), sheet = "TNM")


# load read counts
mqc_kraken <- read_delim(paste0(PARAM$folder, "data/metag/multiqc_metag/mqc_kraken-top-n-plot_Species.txt"), 
                         delim = "\t", escape_double = FALSE, trim_ws = TRUE)
colnames(mqc_kraken) <- c("Filename", "ReadCount (Homo)", "ReadCount (Microbiome)")
# Replace "-kraken2-report" with an empty string in the specified column
mqc_kraken$Filename <- gsub("-kraken2-report", "", mqc_kraken$Filename)
# Replace "-" with "." in the "Filename" column
mqc_kraken$Filename <- gsub("-", ".", mqc_kraken$Filename)
```

```{r}
# there are a lot of duplicates, rmeove duplicate rows
meta2=distinct(meta2)
meta3=distinct(meta3)
# Combine the two dataframes by StudentID
# Pivot the long table into a wider format
# wide_data <- meta2 %>%
#   pivot_wider(names_from = Timepoint, values_from = Note)
df1 <- Metadata_PROMISE %>%
  left_join(meta2, by = c("StudienID", "Timepoint"))
 # left_join(meta3, by = c("StudienID", "Timepoint")) 

df2  <- df1 %>%
  left_join(meta1, by = c("StudienID"))
df3 <- df2 %>%
  full_join(mqc_kraken, by = c("Filename"))
```

Response variable can not be added because so many patients have double coded respoonse that does not make sense.
For e.g. SD (Stable disease) and PD (Progressive disease) for the same patient and same timepoint 
```{r}
meta=df3
meta$`ReadCount (Microbiome)` = as.integer(meta$`ReadCount (Microbiome)`)
meta$ID <- paste0(meta$SampleName, "_", meta$Timepoint, "_", meta$Sequencing_batch, "_",meta$Resubmission )

#recode metadat continuous c=variables
meta <- meta %>%
  mutate(Timepoint = ifelse(Timepoint == "0", "Control", Timepoint)) %>%
  filter(Sequencing_batch %in% c(0, 1, 2, 3)) %>%
  mutate(
    age_factor = as.factor(
      cut(Age, 
          breaks = quantile(Age, na.rm = TRUE), 
          labels = c(1, 2, 3, 4), na.rm = TRUE)
    ),
    lib_size_factor = as.factor(
      cut(`ReadCount (Microbiome)`, 
          breaks = quantile(`ReadCount (Microbiome)`, na.rm = TRUE), 
          labels = c(1, 2, 3, 4), na.rm = TRUE)
    ),
    lib_Homo_factor = as.factor(
      cut(`ReadCount (Homo)`, 
          breaks = quantile(`ReadCount (Homo)`, na.rm = TRUE), 
          labels = c(1, 2, 3, 4), na.rm = TRUE)
    )
  )

```

```{r save files, message=FALSE, echo=FALSE}
# Save Rdata
save(meta, 
     file=paste0(PARAM$folder.Rdata, "metaBatch123.Rdata"))
```
